Moved block and template related classes to their respective folders
