+++ src/Template/TemplateAwareInterface.php
@@ -1,18 +1,20 @@
 <?php
 
-namespace Dhii\Output;
+namespace Dhii\Output\Template;
+
+use Dhii\Output\Template\TemplateInterface;
 
 /**
  * Something that can have a template retrieved from it.
  *
- * @since 0.2
+ * @since [*next-version*]
  */
 interface TemplateAwareInterface
 {
     /**
      * Retrieves the template associated with this instance.
      *
-     * @since 0.2
+     * @since [*next-version*]
      *
      * @return TemplateInterface|null The template, if any.
      */
